% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombe_set1.R
\docType{data}
\name{anscombe_set1}
\alias{anscombe_set1}
\title{Four x-y datasets which have the same traditional statistical properties (mean, variance, correlation, regression line, etc.), yet are quite different.}
\format{
A subset of the \code{anscombe} data frame with 11 observations on 2 variables.
\describe{
\item{x1}{dbl random coordinate}
\item{y1}{dbl random coordinate}
}
}
\source{
\url{https://https://stat.ethz.ch/R-manual/R-patched/library/datasets/html/anscombe.html}
}
\usage{
anscombe_set1
}
\description{
Four x-y datasets which have the same traditional statistical properties (mean, variance, correlation, regression line, etc.), yet are quite different.
}
\keyword{datasets}
