% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datian-package.R
\docType{package}
\name{datian-package}
\alias{datian}
\alias{datian-package}
\title{datian: Author package for storing raw and cleaned data}
\description{
The `datian` package simply stores readily accessible data for my own use.
}
\author{
\strong{Maintainer}: Ian Adams \email{adams.po@gmail.com} (\href{https://orcid.org/0000-0001-5595-8070}{ORCID})

}
\keyword{internal}
